+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog for package xsens_mti_ros_node
 
+## 3.0.0 (2016-05-27)
+* Added functionality to detect devices based on Product masks (incl. 1-series and FMT1000 devices)
+* Increased the 'additionalTimeOutOffset' from 6ms to 10ms as default
+
 ## 2.0.2 (2015-11-23)
 * Fix Readme